+++ instrumentation/excon/test/opentelemetry/instrumentation/excon/stable/instrumentation_test.rb
@@ -71,6 +71,26 @@ describe OpenTelemetry::Instrumentation::Excon::Instrumentation do
       _(span.attributes['http.request.method']).must_equal 'GET'
     end
 
+    it 'handles unknown HTTP method' do
+      stub_request(:purge, 'http://example.com/purge').to_return(status: 200)
+
+      Excon.new('http://example.com/purge').request(method: 'PURGE')
+
+      _(exporter.finished_spans.size).must_equal 1
+      _(span.name).must_equal 'HTTP'
+      _(span.attributes['http.request.method']).must_equal '_OTHER'
+      _(span.attributes['http.request.method_original']).must_equal 'PURGE'
+      _(span.attributes['url.scheme']).must_equal 'http'
+      _(span.attributes['http.response.status_code']).must_equal 200
+      _(span.attributes['url.path']).must_equal '/purge'
+      _(span.attributes['url.full']).must_equal 'http://example.com/purge'
+      assert_requested(
+        :purge,
+        'http://example.com/purge',
+        headers: { 'Traceparent' => "00-#{span.hex_trace_id}-#{span.hex_span_id}-01" }
+      )
+    end
+
     it 'after request with failure code' do
       Excon.get('http://example.com/failure')
 
@@ -171,6 +191,19 @@ describe OpenTelemetry::Instrumentation::Excon::Instrumentation do
       end
 
       _(span.attributes['peer.service']).must_equal 'example:custom'
+    end
+
+    describe 'context detachment with non-recording spans' do
+      it 'detaches context when span is not recorded' do
+        with_sampler(OpenTelemetry::SDK::Trace::Samplers::ALWAYS_OFF) do
+          initial_context = OpenTelemetry::Context.current
+          Excon.get('http://example.com/success')
+          final_context = OpenTelemetry::Context.current
+
+          _(final_context).must_equal initial_context
+          _(exporter.finished_spans).must_be_empty
+        end
+      end
     end
   end