+++ instrumentation/http/lib/opentelemetry/instrumentation/http/patches/stable/client.rb
@@ -16,18 +16,20 @@ module OpenTelemetry
             HTTP_STATUS_SUCCESS_RANGE = (100..399)
 
             def perform(req, options)
+              span_data = HttpHelper.span_attrs_for(req.verb)
+
               uri = req.uri
-              request_method = req.verb.to_s.upcase
-              span_name = create_request_span_name(request_method, uri.path)
+              span_name = create_span_name(span_data, uri.path)
 
               attributes = {
-                'http.request.method' => request_method,
+                'http.request.method' => span_data.normalized_method,
                 'url.scheme' => uri.scheme,
                 'url.path' => uri.path,
                 'url.full' => "#{uri.scheme}://#{uri.host}",
                 'server.address' => uri.host,
                 'server.port' => uri.port
               }
+              attributes['http.request.method_original'] = span_data.original_method if span_data.original_method
               attributes['url.query'] = uri.query unless uri.query.nil?
               attributes.merge!(OpenTelemetry::Common::HTTP::ClientContext.attributes)
 
@@ -53,15 +55,17 @@ module OpenTelemetry
               span.status = OpenTelemetry::Trace::Status.error unless HTTP_STATUS_SUCCESS_RANGE.cover?(status_code)
             end
 
-            def create_request_span_name(request_method, request_path)
+            def create_span_name(span_data, request_path)
+              default_span_name = span_data.span_name
+
               if (implementation = config[:span_name_formatter])
-                updated_span_name = implementation.call(request_method, request_path)
-                updated_span_name.is_a?(String) ? updated_span_name : request_method.to_s
+                updated_span_name = implementation.call(span_data.normalized_method, request_path)
+                updated_span_name.is_a?(String) ? updated_span_name : default_span_name
               else
-                request_method.to_s
+                default_span_name
               end
             rescue StandardError
-              request_method.to_s
+              default_span_name
             end
 
             def tracer