+++ instrumentation/anthropic/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release History: opentelemetry-instrumentation-anthropic
 
+### v0.3.0 / 2025-10-22
+
+* BREAKING CHANGE: Min Ruby Version 3.2
+
+* ADDED: Min Ruby Version 3.2
+
 ### v0.2.1 / 2025-09-30
 
 * FIXED: Min OTel Ruby API 1.7