+++ propagator/google_cloud_trace_context/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release History: opentelemetry-propagator-google_cloud_trace_context
 
+### v0.3.0 / 2025-10-22
+
+* BREAKING CHANGE: Min Ruby Version 3.2
+
+* ADDED: Min Ruby Version 3.2
+
 ### v0.2.0 / 2025-09-30
 
 * ADDED: Bump minimum API Version to 1.7